Progress towards transitioning to Location Graph
