+++ Anthology/Anthology.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="Data\" />
+    <Folder Include="Models\MapManager\" />
   </ItemGroup>
 
   <ItemGroup>
