+++ Anthology/Data/Paths.json
@@ -1,5 +1,5 @@
 {
-  "Locations": "Data\\Locations.json",
+  "Locations": "Data\\Locations2.json",
   "Actions": "Data\\Actions.json",
   "Agents": "Data\\Agents.json"
 }
