+++ Anthology/Models/Deprecated/LocationNode.cs
@@ -1,30 +1,23 @@
-﻿using System.Text.Json.Serialization;
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Numerics;
+using System.Runtime.CompilerServices;
+using System.Text.Json.Serialization;
 
-namespace Anthology.Models
+namespace Anthology.Models.MapManager
 {
-    public class SimLocation
+    public class LocationNode
     {
-        /** optional name of the location. Eg. Restaurant, Home, Movie Theatre, etc */
         public string Name { get; set; } = string.Empty;
 
-        /** x-coordinate of the location */
-        public int X { get; set; }
+        public Vector2 Position { get; set; }
 
-        /** y-coordinate of the location */
-        public int Y { get; set; }
+        public HashSet<string> Tags { get; set; } = new();
 
-        /** optional set of tags associated with the location. Eg. Restaurant could have 'food', 'delivery' as tags */
-        public HashSet<string> Tags { get; set; } = new HashSet<String>();
+        public Dictionary<string, float> ConnectedLocations { get; set; } = new();
 
-        /** set of agents at the location */
         [JsonIgnore]
-        public HashSet<string> AgentsPresent { get; set; } = new HashSet<string>();
-
-        /** returns true if the specified agent is at this location */
-        public bool IsAgentHere(Agent npc)
-        {
-            return AgentsPresent.Contains(npc.Name);
-        }
+        public HashSet<string> AgentsPresent { get; set; } = new();
 
         /** checks if this location satisfies all of the passed location requirements */
         public bool SatisfiesRequirements(RLocation reqs)
@@ -108,5 +101,53 @@ namespace Anthology.Models
             }
             return relationshipsPresent.IsSubsetOf(relationshipsHere);
         }
+
+        /*[JsonIgnore]
+        public Dictionary<string, Tuple<LocationNode, float>> ClosestLocationByTag { get; set; } = new();
+
+        public void UpdateClosestLocationsByTags(HashSet<string> tags)
+        {
+            ClosestLocationByTag.Clear();
+            foreach (string tag in tags)
+            {
+                ClosestLocationByTag.Add(tag, FindClosestLocationByTag(tag));
+            }
+        }
+
+        private Tuple<LocationNode, float> FindClosestLocationByTag(string tag)
+        {
+            Dictionary<string, float> matches = new();
+            HashSet<string> searched = new();
+            float shortest = float.MaxValue;
+            ClosestLocationHelper(tag, ref shortest, matches, searched);
+
+            foreach (KeyValuePair<string, float> kvp in matches)
+            {
+                if (kvp.Value == shortest)
+                {
+                    return new Tuple<LocationNode, float>(LocationManager.LocationsByName[kvp.Key], shortest);
+                }
+            }
+
+            throw new ArgumentException("Unable to reach a location with tag " + tag + " from " + Name);
+        }
+
+        private void ClosestLocationHelper(string tag, ref float shortest, Dictionary<string, float> matches, HashSet<string> searched, float csf = 0)
+        {
+            searched.Add(this.Name);
+            if (Tags.Contains(tag))
+            {
+                matches.Add(this.Name, csf);
+                if (csf < shortest)
+                    shortest = csf;
+                return;
+            }
+            
+            foreach (KeyValuePair<string, float> kvp in ConnectedLocations)
+            {
+                if (searched.Contains(kvp.Key) || csf + kvp.Value > shortest) continue;
+                LocationManager.LocationsByName[kvp.Key].ClosestLocationHelper(tag, ref shortest, matches, searched, csf + kvp.Value);
+            }
+        }*/
     }
 }
