+++ SimManagerUnitTest/Data/Locations2.json
@@ -1,0 +1,78 @@
+[
+  {
+    "Name": "Dining Hall",
+    "X": 5,
+    "Y": 5,
+    "Tags": [
+      "restaurant"
+    ],
+    "Connections": {
+      "Computer Lab": 1,
+      "Dorm": 5
+    }
+  },
+  {
+    "Name": "Computer Lab",
+    "X": 4,
+    "Y": 5,
+    "Tags": [
+      "has:computer"
+    ],
+    "Connections": {
+      "Dining Hall": 1,
+      "Dorm": 4
+    }
+  },
+  {
+    "Name": "Dorm",
+    "X": 3,
+    "Y": 2,
+    "Tags": [
+      "dorm",
+      "has:bed",
+      "has:computer"
+    ],
+    "Connections": {
+      "Dining Hall": 5,
+      "Computer Lab": 4,
+      "Greenway": 2
+    }
+  },
+  {
+    "Name": "Greenway",
+    "X": 1,
+    "Y": 2,
+    "Tags": [
+      "outdoor"
+    ],
+    "Connections": {
+      "Dorm": 2,
+      "Math Hall": 1,
+      "Physics Hall": 1
+    }
+  },
+  {
+    "Name": "Math Hall",
+    "X": 1,
+    "Y": 1,
+    "Tags": [
+      "classroom",
+      "math"
+    ],
+    "Connections": {
+      "Greenway": 1
+    }
+  },
+  {
+    "Name": "Physics Hall",
+    "X": 1,
+    "Y": 3,
+    "Tags": [
+      "classroom",
+      "physics"
+    ],
+    "Connections": {
+      "Greenway": 1
+    }
+  }
+]
